+++ QSymbolicsBase/CHANGELOG.md
@@ -1,6 +1,11 @@
 # News
 
-## v0.1.1 - dev
+## v0.1.2
+
+- Pretty printing with Latexify.
+- Implement bras.
+
+## v0.1.1
 
 - Bosonic states basis and common states and operators.
 - Documentation and more symbol names available for the qubit base states.